# openapi.yaml
openapi: 3.0.3
info:
  title: Awesome Pizza API
  version: 1.0.0
  description: |
    A RESTful API for managing the daily pizza menu and customer orders,
    built with Express.js and TypeScript.

    The API uses an **in-memory database**, so all data resets when the server
    restarts. Three sample orders (`order-001`, `order-002`, `order-003`) are
    seeded on startup.

    ## Response envelope
    Every response follows a standard envelope:

    - **Success:** `{ "success": true, "data": <payload>, "message": "..." }`
    - **Error:** `{ "success": false, "error": "<type>", "message": "..." }`

    ## Authentication
    Two endpoints (`/api/protected` and `/api/admin`) require a Bearer token.
    The token is a Base64-encoded JSON payload describing a role, e.g.
    `Buffer.from(JSON.stringify({ role: 'admin' })).toString('base64')`
    produces `eyJyb2xlIjoiYWRtaW4ifQ==`.
  license:
    name: ISC
  contact:
    name: awesome-pizza
    url: https://github.com/alexhddev/awesome-pizza

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Daily Menu
    description: The list of pizzas available today
  - name: Orders
    description: Create, read and update pizza orders
  - name: Auth
    description: Token-protected demo endpoints

paths:
  /api/daily-menu:
    get:
      tags: [Daily Menu]
      summary: Get the daily menu
      description: Returns the current list of pizzas available for the day.
      operationId: getDailyMenu
      responses:
        '200':
          description: Daily menu retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyMenuResponse'
              example:
                success: true
                message: Daily menu retrieved successfully
                data:
                  - name: Margherita Pizza
                    description: Classic pizza with fresh tomatoes, mozzarella cheese, and basil
                    imageUrl: assets/origs/margherita.png
                  - name: Pepperoni Pizza
                    description: Traditional pizza topped with pepperoni and mozzarella cheese
                    imageUrl: assets/origs/pepperoni.png
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/orders:
    post:
      tags: [Orders]
      summary: Create a new order
      description: |
        Creates a new pizza order. The `status` is always set to `RECEIVED`
        and the `id` is generated by the server; both fields are ignored if
        supplied in the request body.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              sender: Alice Johnson
              contents:
                - name: Margherita Pizza
                  quantity: 2
                - name: BBQ Chicken Pizza
                  quantity: 1
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                success: true
                message: Order created successfully
                data:
                  id: order-1698123456789-123
                  sender: Alice Johnson
                  status: RECEIVED
                  contents:
                    - name: Margherita Pizza
                      quantity: 2
                    - name: BBQ Chicken Pizza
                      quantity: 1
        '400':
          description: |
            Validation error. Triggered when:
            - `sender` is missing or not a non-empty string, or
            - `contents` is missing, not an array, or empty, or
            - any item is missing a valid `name`, or
            - any item is missing a valid `quantity` (positive number).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Bad request
                message: Order sender is required and must be a non-empty string
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order by ID
      operationId: getOrderById
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                success: true
                message: Order retrieved successfully
                data:
                  id: order-001
                  sender: John Doe
                  status: RECEIVED
                  contents:
                    - name: Margherita Pizza
                      quantity: 2
                    - name: Pepperoni Pizza
                      quantity: 1
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Orders]
      summary: Update an existing order
      description: |
        Updates one or more fields of an existing order. All body fields are
        optional; only the fields provided are changed.
      operationId: updateOrderById
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
            example:
              sender: John Doe Updated
              status: DELIVERING
              contents:
                - name: Margherita Pizza
                  quantity: 3
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                success: true
                message: Order updated successfully
                data:
                  id: order-001
                  sender: John Doe Updated
                  status: DELIVERING
                  contents:
                    - name: Margherita Pizza
                      quantity: 3
        '400':
          description: |
            Validation error. Triggered when a provided field is invalid:
            - `sender` is not a non-empty string, or
            - `status` is not one of `RECEIVED`, `DELIVERING`, `DELIVERED`, `CANCELED`, or
            - `contents` is not a non-empty array, or an item has an invalid `name`/`quantity`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Bad request
                message: 'Status must be one of: RECEIVED, DELIVERING, DELIVERED, CANCELED'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/protected:
    get:
      tags: [Auth]
      summary: Protected route (token required)
      description: |
        Returns 200 only when a valid Bearer token is supplied. The token must
        exactly equal the hardcoded admin token `eyJyb2xlIjoiYWRtaW4ifQ==`.
      operationId: getProtected
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Access granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                message: Access granted to protected route
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Unauthorized
                message: Valid token required

  /api/admin:
    get:
      tags: [Auth]
      summary: Admin route (role check)
      description: |
        Decodes the Base64 Bearer token as JSON and grants access only when
        `role` equals `admin`.
      operationId: getAdmin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Access granted for admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                message: Welcome, admin!
        '401':
          description: Token missing or not valid Base64 JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingToken:
                  value:
                    success: false
                    error: Unauthorized
                    message: Token required
                invalidFormat:
                  value:
                    success: false
                    error: Unauthorized
                    message: Invalid token format
        '403':
          description: Token decoded but role is not admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Forbidden
                message: Not authorised

components:
  parameters:
    OrderId:
      name: id
      in: path
      required: true
      description: Unique order identifier
      schema:
        type: string
      example: order-001

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Base64-encoded JSON token. Use `eyJyb2xlIjoiYWRtaW4ifQ==` (role: admin)
        to access both protected and admin routes.

  schemas:
    MenuEntry:
      type: object
      required: [name, description, imageUrl]
      properties:
        name:
          type: string
          example: Margherita Pizza
        description:
          type: string
          example: Classic pizza with fresh tomatoes, mozzarella cheese, and basil
        imageUrl:
          type: string
          example: assets/origs/margherita.png

    OrderItem:
      type: object
      required: [name, quantity]
      properties:
        name:
          type: string
          description: Name of the pizza
          example: Margherita Pizza
        quantity:
          type: integer
          minimum: 1
          description: Number of this pizza ordered (must be a positive number)
          example: 2

    OrderStatus:
      type: string
      description: |
        Lifecycle status of an order:
        - `RECEIVED`   - received and being prepared
        - `DELIVERING` - out for delivery
        - `DELIVERED`  - successfully delivered
        - `CANCELED`   - canceled
      enum:
        - RECEIVED
        - DELIVERING
        - DELIVERED
        - CANCELED

    Order:
      type: object
      required: [id, sender, status, contents]
      properties:
        id:
          type: string
          example: order-001
        sender:
          type: string
          description: Name of the customer who placed the order
          example: John Doe
        status:
          $ref: '#/components/schemas/OrderStatus'
        contents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'

    CreateOrderRequest:
      type: object
      required: [sender, contents]
      properties:
        sender:
          type: string
          description: Name of the customer placing the order
          example: Alice Johnson
        contents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'

    UpdateOrderRequest:
      type: object
      description: All fields are optional; only provided fields are updated.
      minProperties: 1
      properties:
        sender:
          type: string
          example: John Smith
        status:
          $ref: '#/components/schemas/OrderStatus'
        contents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'

    DailyMenuResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/MenuEntry'
        message:
          type: string
          example: Daily menu retrieved successfully

    OrderResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Order'
        message:
          type: string
          example: Order retrieved successfully

    MessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Access granted to protected route

    ErrorResponse:
      type: object
      required: [success, error, message]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Short error type
          example: Not found
        message:
          type: string
          description: Human-readable error detail
          example: "Order with ID 'order-999' not found"

  responses:
    OrderNotFound:
      description: The order with the given ID does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not found
            message: "Order with ID 'order-999' not found"

    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal server error
            message: Failed to retrieve order